52846: It is impossible to select edge in OB as reversed edge for hypothesis 52826: Radial Quadrangle algorithm meshes a quad face from which a circle was cut from but it must fail 52227: Group is not shown during edition 52780: Wrong mesh after STL re-import of a polygonal element 52825: Sew Borders does not care of segment adjacent to split faces, thus segments become non-conformal 52824: Sew Free Borders sews borders as conformal thus producing very skewed quads instead of polygons

+++ src/SMDS/SMDS_Mesh.cxx
@@ -2484,7 +2484,7 @@ const SMDS_MeshElement* SMDS_Mesh::FindElement (const vector<const SMDS_MeshNode
       int nbNodesToCheck = noMedium ? e->NbCornerNodes() : e->NbNodes();
       if ( nbNodesToCheck == nodes.size() )
       {
-        for ( int i = 1; e && i < nodes.size(); ++ i )
+        for ( size_t i = 1; e && i < nodes.size(); ++i )
         {
           int nodeIndex = e->GetNodeIndex( nodes[ i ]);
           if ( nodeIndex < 0 || nodeIndex >= nbNodesToCheck )
